fix: useUpdateLocalCode 분리
